refactor(issue): move close issue to separate file
